--- src/pages/LandingPage/component/FeatureCategoryCard.jsx
@@ -1,15 +1,0 @@
-// const FeatureCategoryCard = ({itemId, videoThumbnail, title, }) => {
-//   return (
-//     <div className="feature-card-container" key={itemId} onClick={() => categoryClickHandler(itemId)}>
-//       <img
-//         src={videoThumbnail}
-//         alt="cateory thumbnail"
-//       />
-//       <div>
-//         <p> {title}</p>
-//       </div>
-//     </div>
-//   );
-// };
-
-// export { FeatureCategoryCard };
