--- src/pages/LandingPage/landing-page.css
@@ -1,104 +1,0 @@
-.featured-category {
-  margin-bottom: 3rem;
-}
-
-.featured-category h1{
-    text-align: start;
-    padding-left: 4rem;
-}
-.featured-category .featured-category-banner {
-  width: 100%;
-  display: flex;
-  justify-content: left;
-  align-items: center;
-  flex-wrap: nowrap;
-  overflow-x: auto;
-  scroll-behavior: smooth;
-  scrollbar-width: none;
-  padding-left: 2rem;
-}
-
-.featured-category-banner::-webkit-scrollbar {
-  display: none;
-}
-
-/* featured category card */
-
-.feature-card-container {
-  min-width: 30%;
-  height: 25rem;
-  border-radius: 10px;
-  margin: auto;
-  margin-left: 2.5rem;
-  box-shadow: rgba(0, 0, 0, 0.16) 0px 1px 4px;
-  position: relative;
-  cursor: pointer;
-}
-
-.feature-card-container img {
-  height: 100%;
-  width: 100%;
-  border-radius: 10px;
-  object-fit: cover;
-}
-
-.feature-card-container div {
-  display: none;
-}
-
-.feature-card-container:hover div {
-  display: block;
-  position: absolute;
-  bottom: 0;
-  width: 100%;
-  height: 100%;
-  background-color: #7c7c7c;
-  border-radius: 10px;
-  transition-duration: 4s;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.feature-card-container p {
-  text-align: center;
-  font-size: 3rem;
-  padding: 2.6rem;
-  color: var(--white-color);
-}
-
-
-@media screen and (max-width: 1200px) {
-    .feature-card-container {
-        min-width: 40%;
-        height: 23rem;
-    }
-}
-
-@media screen and (max-width: 1000px) {
-    .feature-card-container {
-        min-width: 50%;
-        height: 22rem;
-    }
-}
-
-
-@media screen and (max-width: 800px) {
-    .feature-card-container {
-        min-width: 50%;
-        height: 20rem;
-    }
-}
-
-@media screen and (max-width: 600px) {
-    .feature-card-container {
-        min-width: 90%;
-        margin-left: .5rem;
-    }
-
-    .featured-category-banner {
-        width: 90%;
-        justify-content: space-between;
-        gap: 2rem;
-    }
-}
