+++ src/pages/VideoListing/videoListing.css
@@ -5,16 +5,28 @@
 .video-listing{
     min-height: calc(100vh - 5rem);
 }
-.video-category{
-    padding:  2rem ;
+
+.video-listing .featured-category{
+    /* border: 1px solid black; */
+    padding-left: 2rem;
 }
 
-.video-listing .video-category .btn{
-    box-shadow: none;
-    padding: .5rem 1rem;
-    border-radius: 10px;
-    cursor: pointer;
-    background-color: var(--primary);
+.video-listing .featured-category .btn{
+    background-color: var(--light-color);
+    color: var(--black-color);
+    border-radius: 2rem;
+    font-size: 1.5rem;
+    margin: 0;
+    margin-right: 1.5rem;
+}
+
+.video-listing .featured-category .btn:active{
+ background-color: var(--ligtest-color);
+}
+
+.video-listing .featured-category .active{
+    background-color: var(--black-color);
+    color: var(--white-color);
 }
 
 .video-listin-flex{
@@ -27,4 +39,6 @@
   flex-wrap: wrap;
 }
 
-
+.featured-category{
+    margin-top: 2rem;
+}
